Refactor: Rename and move the validator class
